ai-service/requirements.txt
fastapi
uvicorn
python-dotenv
pytesseract
pillow
transformers
torch
numpy
scikit-learn
pydantic
requests
